logical changes in bash ansible script

+++ scripts/ansible_install.sh
@@ -1,12 +1,14 @@
 #!/bin/bash
 # a bash script to install ansible in various linux machines
 
-if [[ -f /usr/bin/apt ]]; then
-	apt-get update -y
-	apt-get install -y ansible
-elif [[ -f /usr/bin/yum ]]; then
-	yum install -y epel-release
-	yum install -y ansible
-elif [[ -f /usr/bin/zypper ]]; then
-	zypper install -y ansible
+if [[ ! -f /usr/bin/ansible ]]; then
+	if [[ -f /usr/bin/apt ]]; then
+		apt-get update -y
+		apt-get install -y ansible
+	elif [[ -f /usr/bin/yum ]]; then
+		yum install -y epel-release
+		yum install -y ansible
+	elif [[ -f /usr/bin/zypper ]]; then
+		zypper install -y ansible
+	fi
 fi
